+++ sdk/trace/src/main/java/io/opentelemetry/sdk/trace/OnEndSpanProcessor.java
@@ -27,7 +27,7 @@ public final class OnEndSpanProcessor implements SpanProcessor {
 
   @Override
   public boolean isEndRequired() {
-    return false;
+    return true;
   }
 
   @Override
