+++ sdk/trace/src/test/java/io/opentelemetry/sdk/trace/OnStartSpanProcessorTest.java
@@ -5,14 +5,14 @@
 
 package io.opentelemetry.sdk.trace;
 
-import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 
 import io.opentelemetry.context.Context;
 import java.util.concurrent.atomic.AtomicReference;
 import org.junit.jupiter.api.Test;
 
-class SpanProcessorTest {
+class OnStartSpanProcessorTest {
 
   @Test
   void startOnly() {
@@ -22,7 +22,7 @@ class SpanProcessorTest {
     ReadWriteSpan inputSpan = mock(ReadWriteSpan.class);
 
     SpanProcessor processor =
-        SpanProcessor.startOnly(
+        OnStartSpanProcessor.create(
             (ctx, span) -> {
               seenContext.set(ctx);
               seenSpan.set(span);
@@ -32,19 +32,6 @@ class SpanProcessorTest {
     assertThat(processor.isEndRequired()).isFalse();
     processor.onStart(context, inputSpan);
     assertThat(seenContext.get()).isSameAs(context);
-    assertThat(seenSpan.get()).isSameAs(inputSpan);
-  }
-
-  @Test
-  void endOnly() {
-    AtomicReference<ReadableSpan> seenSpan = new AtomicReference<>();
-    ReadWriteSpan inputSpan = mock(ReadWriteSpan.class);
-
-    SpanProcessor processor = SpanProcessor.endOnly(seenSpan::set);
-
-    assertThat(processor.isStartRequired()).isFalse();
-    assertThat(processor.isEndRequired()).isTrue();
-    processor.onEnd(inputSpan);
     assertThat(seenSpan.get()).isSameAs(inputSpan);
   }
 }
